Define a module to store useful code used in main pipelines
