+++ train_test.py
@@ -18,215 +18,37 @@ import torch.nn as nn
 import torch.optim
 from tqdm import tqdm
 from collections import defaultdict
+
+
 ## 2. Configuration de l'environnement
 ### 2.1 Test de la disponibilité d'un GPU compatible avec Pytorch
-if torch.cuda.is_available():
-    print("Un GPU compatible CUDA a été détecté.")
-else:
-    print("Aucun GPU compatible CUDA n'a été détecté. Il vaut certainement mieux utiliser une instance de Google Colab")
 
 ### 2.2 Définition de classes utilitaires
-class DataReader:
-    """
-    A helper class to read image and mask data from specified paths.
-    """
-    @staticmethod
-    def read_data(image_path, mask_path=None):
-        """
-        Reads an image and its corresponding mask from the given file paths.
-        Args:
-            image_path (str): Path to the image file.
-            mask_path (str): Path to the mask file.
-        Returns:
-            tuple: A tuple containing the original image and mask as numpy arrays.
-        """
-        image = cv2.imread(image_path)
-        original_image = cv2.cvtColor(image, cv2.COLOR_BGR2RGB)
-        if mask_path :
-          original_mask = cv2.imread(mask_path, cv2.IMREAD_GRAYSCALE)
-          return original_image, original_mask
-        else:
-          return original_image, None
 
 
-def preprocess_mask(mask):
-    mask = mask.astype(np.float32)
-    mask[mask == 255.0] = 1.0
-    return mask
 
 
-class Visualizer:
-    """
-    A helper class for visualizing images and masks.
-    """
-    @staticmethod
-    def visualize(image, mask, original_image=None, original_mask=None, title=None):
-        """
-        Visualizes the original and transformed images and masks.
-        Args:
-            image (numpy.ndarray): Transformed image to be visualized.
-            mask (numpy.ndarray): Transformed mask to be visualized.
-            original_image (numpy.ndarray, optional): Original image for comparison.
-            original_mask (numpy.ndarray, optional): Original mask for comparison.
-            title (str, optional): Title for the visualization plot.
-        """
-        fontsize = 18
-        if original_image is None and original_mask is None:
-            f, ax = plt.subplots(2, 1, figsize=(8, 8))
-            ax[0].imshow(image)
-            ax[1].imshow(mask)
-        else:
-            f, ax = plt.subplots(2, 2, figsize=(8, 8))
-            f.suptitle(title, fontsize=fontsize)
-            ax[0, 0].imshow(original_image)
-            ax[0, 0].set_title('Original image', fontsize=fontsize)
-            ax[1, 0].imshow(original_mask)
-            ax[1, 0].set_title('Original mask', fontsize=fontsize)
-            ax[0, 1].imshow(image)
-            ax[0, 1].set_title('Transformed image', fontsize=fontsize)
-            ax[1, 1].imshow(mask)
-            ax[1, 1].set_title('Transformed mask', fontsize=fontsize)
 
 
-    @staticmethod
-    def visualize_batch(data_loader):
-        """
-        Visualizes the first three images and masks from the given data loader.
-
-        Args:
-            data_loader (DataLoader): A PyTorch DataLoader object containing images and masks.
-        """
-        fig, axs = plt.subplots(3, 2, figsize=(10, 15))  # 3 rows, 2 columns
-        mean = torch.tensor([0.485, 0.456, 0.406])
-        std = torch.tensor([0.229, 0.224, 0.225])
-        mean = mean[:, None, None]
-        std = std[:, None, None]
-        for i, (images, masks) in enumerate(data_loader):
-            if i == 3:  # Only show first 3 images and masks
-                break
-
-            axs[i, 0].imshow((torch.clamp((images[i]*std+mean).permute(1, 2, 0), 0, 1)*255).numpy().astype(np.uint8))  # Change CxHxW to HxWxC
-            axs[i, 0].set_title(f'Image {i}')
-            axs[i, 0].axis('off')
-
-            axs[i, 1].imshow(masks[i], cmap='gray')
-            axs[i, 1].set_title(f'Mask {i} & Pourcentage of parking : {masks[i][masks[i]==1.0].shape[0]/(256*256)}')
-            axs[i, 1].axis('off')
-
-        plt.tight_layout()
-        plt.show()
-
-    @staticmethod
-    def display_image_grid(images_filenames, images_directory, masks_directory, predicted_masks=None):
-        cols = 3 if predicted_masks else 2
-        rows = len(images_filenames)
-        figure, ax = plt.subplots(nrows=rows, ncols=cols, figsize=(10, 24))
-        for i, image_filename in enumerate(images_filenames):
-            image = cv2.imread(os.path.join(images_directory, image_filename))
-            image = cv2.cvtColor(image, cv2.COLOR_BGR2RGB)
-            mask = cv2.imread(os.path.join(masks_directory, image_filename.replace(".png", "_mask.png")), cv2.IMREAD_GRAYSCALE)
-            mask = preprocess_mask(mask)
-            ax[i, 0].imshow(image)
-            ax[i, 1].imshow(mask, interpolation="nearest")
-
-            ax[i, 0].set_title("Image")
-            ax[i, 1].set_title("Ground truth mask")
-
-            ax[i, 0].set_axis_off()
-            ax[i, 1].set_axis_off()
-
-            if predicted_masks:
-                predicted_mask = predicted_masks[i]
-                ax[i, 2].imshow(predicted_mask, interpolation="nearest")
-                ax[i, 2].set_title("Predicted mask")
-                ax[i, 2].set_axis_off()
-        plt.tight_layout()
-        plt.show()
 
 ### 2.3 Définition des classes afin de créer les datasets
-class FreeParkingPlacesDataset(Dataset):
-    def __init__(self, images_directory, masks_directory, transform=None):
-        self.images_directory = images_directory
-        self.masks_directory = masks_directory
-        self.images_filenames = os.listdir(images_directory)
-        self.transform = transform
-
-    def __len__(self):
-        return len(self.images_filenames)
-
-    def __getitem__(self, idx):
-        image_filename = self.images_filenames[idx]
-        image_path = os.path.join(self.images_directory, image_filename)
-        mask_path = os.path.join(self.masks_directory, image_filename.replace(".png", '_mask.png'))
-        image, mask = DataReader.read_data(image_path, mask_path)
-        mask = self.preprocess_mask(mask)
-        if self.transform is not None:
-            transformed = self.transform(image=image, mask=mask)
-            image = transformed["image"]
-            mask = transformed["mask"]
-        return image, mask
-
-    def preprocess_mask(self, mask):
-        mask = mask.astype(np.float32)
-        mask[mask == 255.0] = 1.0
-        return mask
-class FreeParkingPlacesInferenceDataset(Dataset):
-    def __init__(self, images_directory, transform=None):
-        self.images_directory = images_directory
-        self.images_filenames = os.listdir(images_directory)
-        self.transform = transform
-
-    def __len__(self):
-        return len(self.images_filenames)
-
-    def __getitem__(self, idx):
-        image_filename = self.images_filenames[idx]
-        image_path = os.path.join(self.images_directory, image_filename)
-        image, _ = DataReader.read_data(image_path)
-        original_size = tuple(image.shape[:2])
-        if self.transform is not None:
-            transformed = self.transform(image=image)
-            image = transformed["image"]
-        return image, original_size
-
-    def preprocess_mask(self, mask):
-        mask = mask.astype(np.float32)
-        mask[mask == 255.0] = 1.0
-        return mask
 ### 2.4 Constants & Helper Entities definitions
-visualizer_batch_size = 4
 visualizer_worker = Visualizer()
 
-# training_image_dir = "./drive/MyDrive/ML_with_Python_Project_Segmentation/data/train/images/"
-# training_mask_dir = "./drive/MyDrive/ML_with_Python_Project_Segmentation/data/train/masks/"
 
-# val_image_dir = "./drive/MyDrive/ML_with_Python_Project_Segmentation/data/val/images/"
-# val_mask_dir = "./drive/MyDrive/ML_with_Python_Project_Segmentation/data/val/masks/"
+# training_image_dir = "../data/train/images/"
+# training_mask_dir = "../data/train/masks/"
 
-# test_image_dir = "./drive/MyDrive/ML_with_Python_Project_Segmentation/data/test/images/"
-# test_mask_dir = "./drive/MyDrive/ML_with_Python_Project_Segmentation/data/test/masks/"
-# test_image_filenames = os.listdir("./drive/MyDrive/ML_with_Python_Project_Segmentation/data/test/images/")
+# val_image_dir = "../data/val/images/"
+# val_mask_dir = "../data/val/masks/"
 
-training_image_dir = "../data/train/images/"
-training_mask_dir = "../data/train/masks/"
+# test_image_dir = "../data/test/images/"
+# test_mask_dir = "../data/test/masks/"
 
-val_image_dir = "../data/val/images/"
-val_mask_dir = "../data/val/masks/"
-
-test_image_dir = "../data/test/images/"
-test_mask_dir = "../data/test/masks/"
 test_image_filenames = os.listdir("../data/test/images/")
 
 
-params = {
-    "model": "UNet11",
-    "device": "cuda" if torch.cuda.is_available() else "cpu",
-    "lr": 0.003,
-    "class_weights": [1.0, 3.0],
-    "batch_size": 16,
-    # "num_workers": 4,
-    "epochs": 25,
-}
+
 ### 2.5 Definition of Training & Validation DataSets
 training_ds = FreeParkingPlacesDataset(
     images_directory=training_image_dir,
